requirements.txt
aiohttp>=3.8.0
aiofiles>=22.0.0
weixin-bot-sdk
qrcode[pil]>=7.4.0
